Removed Movielist component and added the logic into moviescreen file
