+++ src/screens/MovieDetails.js
@@ -10,20 +10,22 @@ const MovieDetails = () => {
       <Text>Title {movie.Title}</Text>
       <Text>Year {movie.Year}</Text>
       <Text>Released {movie.Released}</Text>
-      <Image style={styles.logo}
-        source={{
-          uri: movie?.Poster,
-        }} />
+      {movie?.Poster &&
+        <Image style={styles.logo}
+          source={{
+            uri: movie?.Poster,
+          }} />
+      }
     </View>
   )
 }
 const styles = StyleSheet.create({
   container: {
     padding: 5,
-    margin:15
+    margin: 15
   },
   logo: {
-    margin:10,
+    margin: 10,
     width: 250,
     height: 300,
   },
