+++ src/screens/MovieScreen.js
@@ -1,9 +1,9 @@
-import { View, Text, TextInput, StyleSheet, ScrollView } from 'react-native'
+import { View, Text, TextInput, StyleSheet, ScrollView, TouchableOpacity, ActivityIndicator } from 'react-native'
+
 import React, { useEffect, useState } from 'react'
 import { useDispatch } from 'react-redux';
-import { getMovies } from '../features/movieSlice';
-import MovieList from './MovieList';
-
+import { getMovie, getMovies } from '../features/movieSlice';
+import { useNavigation } from '@react-navigation/native';
 import { useSelector } from 'react-redux'
 const MovieScreen = () => {
     const [name, setName] = useState("spider");
@@ -11,15 +11,33 @@ const MovieScreen = () => {
     useEffect(() => {
         dispatch(getMovies(name));
     }, [])
-    const { moviesList: { Error: error }
+    const { moviesList, errorMessage, isError: { Error: error }
     } = useSelector((state) => ({ ...state.movie }));
 
-    onChange = (event) => {    
-        const {eventCount, target, text} = event.nativeEvent;
+    onChange = (event) => {
+        const { eventCount, target, text } = event.nativeEvent;
         setName(text);
         dispatch(getMovies(text));
-      };
-
+    };
+    const ListOfMovies = () => {
+        return <View>
+            {moviesList?.Search?.map((item, index) => (
+                <View key={index}
+                    style={{ padding: 8, borderBottomColor: 'blue', borderBottomWidth: 0.5 }}>
+                    <TouchableOpacity onPress={() => handleNavigation(item.imdbID)}
+                    >
+                        <Text>{item.Title}</Text>
+                        <Text>{item.Year}</Text>
+                    </TouchableOpacity>
+                </View>
+            ))}
+        </View>;
+    };
+    const navigation = useNavigation();
+    const handleNavigation = (id) => {
+        dispatch(getMovie(id));
+        navigation.navigate('MovieDetails')
+    }
     return (
         <ScrollView>
             <View style={styles.container}>
@@ -28,7 +46,8 @@ const MovieScreen = () => {
                     onChange={this.onChange}
                     value={name} />
                 {error && <Text>{error}</Text>}
-                <MovieList></MovieList>
+                <ListOfMovies />
+                {moviesList.length === 0 && <ActivityIndicator />}
             </View>
         </ScrollView>
     )
